requirements.txt
Django==4.2
djangorestframework==3.15.0
gunicorn==20.1.0

# ML / vision
torch>=2.0.0
torchvision
onnxruntime==1.16.0
timm
facenet-pytorch==2.7.0
opencv-python-headless

# Firebase / GCS
firebase-admin==6.0.1
google-cloud-storage==2.11.0

# utils
numpy
Pillow
python-multipart
python-dotenv

opencv-python-headless>=4.10
